Added Features
There was an issue yesterday while updating the count variable which stores the number of guesses made.I could debug the issue now.So after 3 guesses have been made,I added a Toast which shows that only 3 guesses can be made and also disabled the GUESS button. Now,after 3 guesses,the correct answer also gets displayed in the middle of the screen.